requirements.txt
Flask==2.3.3
Flask-Mail==0.9.1
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.31.0 